subreddit2/requirements.txt
requests==2.32.3 
beautifulsoup4==4.12.3 
html5lib==1.1 
selenium==4.24.0
python-dotenv==1.0.1
praw
mysql-connector-python
pandas
undetected-chromedriver
